powerpc/perf: Set cpumode flags using sample address for Power10 and 11
Currently in some cases, when the sampled instruction address register
latches to a specific address during sampling, there is an inconsistency
in the privilege bits captured in the sampled event register.
For example, a snippet from the perf report on a power10 system is:
Overhead  Address             Command       Shared Object      Symbol
........  ..................  ............  .................  .......................
     2.41%  0x7fff9f94a02c      null_syscall  [unknown]          [k] 0x00007fff9f94a02c
     2.20%  0x7fff9f94a02c      null_syscall  libc.so.6          [.] syscall

perf_get_misc_flags() function looks at the privilege bits to return
the corresponding flags to be used for the address symbol and these
privilege bit details are read from the sampled event register. In the
above snippet, address "0x00007fff9f94a02c" is shown as "k" (kernel) due
to the inconsistent privilege bits captured in the sampled event register.
To address this case, the proposed fix is to additionally check whether the
sampled address is in the kernel area. Also, add a flag PPMU_P10_P11 to
identify Power10/11 pmu.
As of commit d9f7088dd6d8 ("powerpc/perf: Use the address from SIAR register
to set cpumode flags"), the sample address is used to set the cpumode flags
for marked events in Power10 DD1. Instead, consider it as an additional check
for Power10/11.

Signed-off-by: Anjali K <anjalik@linux.ibm.com>